+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/sta/model/core/IdLong.java
@@ -77,7 +77,7 @@ public class IdLong implements Id {
     }
 
     @Override
-    public Object getValue() {
+    public Long getValue() {
         return value;
     }