+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/sta/model/core/IdString.java
@@ -73,7 +73,7 @@ public class IdString implements Id {
     }
 
     @Override
-    public Object getValue() {
+    public String getValue() {
         return value;
     }