+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/sta/settings/PersistenceSettings.java
@@ -32,10 +32,12 @@ public class PersistenceSettings {
     private static final String TAG_IMPLEMENTATION_CLASS = "persistenceManagerImplementationClass";
     private static final String DEFAULT_IMPLEMENTATION_CLASS = "de.fraunhofer.iosb.ilt.sta.persistence.postgres.longid.PostgresPersistenceManagerLong";
     private static final String TAG_ALWAYS_ORDERBY_ID = "alwaysOrderbyId";
+    private static final String TAG_ID_GENERATION_MODE = "idGenerationMode";
 
     private static final List<String> ALL_PROPERTIES = Arrays.asList(
             TAG_IMPLEMENTATION_CLASS,
-            TAG_ALWAYS_ORDERBY_ID
+            TAG_ALWAYS_ORDERBY_ID,
+            TAG_ID_GENERATION_MODE
     );
 
     /**
@@ -43,6 +45,7 @@ public class PersistenceSettings {
      */
     private String persistenceManagerImplementationClass;
     private boolean alwaysOrderbyId = true;
+    private String idGenerationMode = "ServerGeneratedOnly";
     /**
      * Extension point for implementation specific settings
      */
@@ -58,6 +61,7 @@ public class PersistenceSettings {
     private void init(Settings settings) {
         persistenceManagerImplementationClass = settings.get(TAG_IMPLEMENTATION_CLASS, DEFAULT_IMPLEMENTATION_CLASS);
         alwaysOrderbyId = settings.getBoolean(TAG_ALWAYS_ORDERBY_ID, alwaysOrderbyId);
+        idGenerationMode = settings.get(TAG_ID_GENERATION_MODE, idGenerationMode);
         customSettings = settings;
     }
 
@@ -71,5 +75,9 @@ public class PersistenceSettings {
 
     public Settings getCustomSettings() {
         return customSettings;
+    }
+    
+    public String getIdGenerationMode() {
+        return idGenerationMode;
     }
 }