+++ FROST-Server.SQL.PGUuid/src/main/java/de/fraunhofer/iosb/ilt/sta/persistence/postgres/uuidid/EntityInserter.java
@@ -21,6 +21,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.dml.StoreClause;
+import com.querydsl.core.types.Path;
 import com.querydsl.core.types.dsl.DateTimePath;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.StringPath;
@@ -127,6 +128,8 @@ public class EntityInserter {
         insert.set(qd.obsPropertyId, (UUID) op.getId().getValue());
         insert.set(qd.sensorId, (UUID) s.getId().getValue());
         insert.set(qd.thingId, (UUID) t.getId().getValue());
+
+        insertUserDefinedId(insert, qd.id, ds);
 
         UUID datastreamId = insert.executeWithKey(qd.id);
         LOGGER.debug("Inserted datastream. Created id = {}.", datastreamId);
@@ -260,6 +263,8 @@ public class EntityInserter {
 
         insert.set(qd.sensorId, (UUID) s.getId().getValue());
         insert.set(qd.thingId, (UUID) t.getId().getValue());
+
+        insertUserDefinedId(insert, qd.id, ds);
 
         UUID multiDatastreamId = insert.executeWithKey(qd.id);
         LOGGER.debug("Inserted multiDatastream. Created id = {}.", multiDatastreamId);
@@ -429,6 +434,8 @@ public class EntityInserter {
         insert.set(qfoi.encodingType, encodingType);
         insertGeometry(insert, qfoi.feature, qfoi.geom, encodingType, foi.getFeature());
 
+        insertUserDefinedId(insert, qfoi.id, foi);
+
         UUID generatedId = insert.executeWithKey(qfoi.id);
         LOGGER.debug("Inserted FeatureOfInterest. Created id = {}.", generatedId);
         foi.setId(new UuidId(generatedId));
@@ -589,6 +596,8 @@ public class EntityInserter {
         insert.set(qhl.time, new Timestamp(h.getTime().getDateTime().getMillis()));
         insert.set(qhl.thingId, (UUID) h.getThing().getId().getValue());
 
+        insertUserDefinedId(insert, qhl.id, h);
+
         UUID generatedId = insert.executeWithKey(qhl.id);
         LOGGER.debug("Inserted HistoricalLocation. Created id = {}.", generatedId);
         h.setId(new UuidId(generatedId));
@@ -667,6 +676,8 @@ public class EntityInserter {
         insert.set(ql.encodingType, encodingType);
         insertGeometry(insert, ql.location, ql.geom, encodingType, l.getLocation());
 
+        insertUserDefinedId(insert, ql.id, l);
+
         UUID locationId = insert.executeWithKey(ql.id);
         LOGGER.debug("Inserted Location. Created id = {}.", locationId);
         l.setId(new UuidId(locationId));
@@ -694,6 +705,7 @@ public class EntityInserter {
             insert = qFactory.insert(qhl);
             insert.set(qhl.thingId, thingId);
             insert.set(qhl.time, new Timestamp(Calendar.getInstance().getTimeInMillis()));
+            // TODO: maybe use histLocationId based on locationId
             UUID histLocationId = insert.executeWithKey(qhl.id);
             LOGGER.debug("Created historicalLocation {}", histLocationId);
 
@@ -811,6 +823,7 @@ public class EntityInserter {
             insert = qFactory.insert(qhl);
             insert.set(qhl.thingId, thingId);
             insert.set(qhl.time, new Timestamp(Calendar.getInstance().getTimeInMillis()));
+            // TODO: maybe use histLocationId based on locationId
             UUID histLocationId = insert.executeWithKey(qhl.id);
             LOGGER.debug("Created historicalLocation {}", histLocationId);
 
@@ -901,6 +914,8 @@ public class EntityInserter {
             insert.set(qo.multiDatastreamId, (UUID) mds.getId().getValue());
         }
         insert.set(qo.featureId, (UUID) f.getId().getValue());
+
+        insertUserDefinedId(insert, qo.id, o);
 
         UUID generatedId = insert.executeWithKey(qo.id);
         LOGGER.debug("Inserted Observation. Created id = {}.", generatedId);
@@ -1048,6 +1063,8 @@ public class EntityInserter {
         insert.set(qop.description, op.getDescription());
         insert.set(qop.properties, objectToJson(op.getProperties()));
 
+        insertUserDefinedId(insert, qop.id, op);
+
         UUID generatedId = insert.executeWithKey(qop.id);
         LOGGER.debug("Inserted ObservedProperty. Created id = {}.", generatedId);
         op.setId(new UuidId(generatedId));
@@ -1145,6 +1162,8 @@ public class EntityInserter {
         // TODO: Check metadata serialisation.
         insert.set(qs.metadata, s.getMetadata().toString());
         insert.set(qs.properties, objectToJson(s.getProperties()));
+
+        insertUserDefinedId(insert, qs.id, s);
 
         UUID generatedId = insert.executeWithKey(qs.id);
         LOGGER.debug("Inserted Sensor. Created id = {}.", generatedId);
@@ -1261,6 +1280,8 @@ public class EntityInserter {
         insert.set(qt.description, t.getDescription());
         insert.set(qt.properties, objectToJson(t.getProperties()));
 
+        insertUserDefinedId(insert, qt.id, t);
+
         UUID thingId = insert.executeWithKey(qt.id);
         LOGGER.debug("Inserted Thing. Created id = {}.", thingId);
         t.setId(new UuidId(thingId));
@@ -1286,6 +1307,7 @@ public class EntityInserter {
             insert = qFactory.insert(qhl);
             insert.set(qhl.thingId, thingId);
             insert.set(qhl.time, new Timestamp(Calendar.getInstance().getTimeInMillis()));
+            // TODO: maybe use histLocationId based on locationIds
             UUID histLocationId = insert.executeWithKey(qhl.id);
             LOGGER.debug("Created historicalLocation {}", histLocationId);
 
@@ -1415,6 +1437,7 @@ public class EntityInserter {
                 SQLInsertClause insert = qFactory.insert(qhl);
                 insert.set(qhl.thingId, thingId);
                 insert.set(qhl.time, new Timestamp(Calendar.getInstance().getTimeInMillis()));
+                // TODO: maybe use histLocationId based on locationIds
                 UUID histLocationId = insert.executeWithKey(qhl.id);
                 LOGGER.debug("Created historicalLocation {}", histLocationId);
 
@@ -1429,6 +1452,14 @@ public class EntityInserter {
             }
         }
         return message;
+    }
+
+    private static <T extends StoreClause> void insertUserDefinedId(T clause, Path idPath, Entity entity) {
+        IdGenerationHandlerUuid idhandler = new IdGenerationHandlerUuid(entity);
+        if (idhandler.useClientSuppliedId()) {
+            idhandler.modifyClientSuppliedId();
+            clause.set(idPath, (UUID) idhandler.getIdValue());
+        }
     }
 
     private static <T extends StoreClause> T insertTimeValue(T clause, DateTimePath<Timestamp> startPath, DateTimePath<Timestamp> endPath, TimeValue time) {
@@ -1540,15 +1571,32 @@ public class EntityInserter {
      * complete and can thus not be created.
      */
     private void entityExistsOrCreate(Entity e) throws NoSuchEntityException, IncompleteEntityException {
-        if (e != null && e.getId() == null) {
+        if (e == null) {
+            throw new NoSuchEntityException("No entity!");
+        }
+
+        if (e.getId() == null) {
             e.complete();
+            // no id but complete -> create
             pm.insert(e);
-        } else if (e == null || !entityExists(e)) {
-            if (e == null) {
-                throw new NoSuchEntityException("No entity!");
-            }
+            return;
+        }
+
+        if (entityExists(e)) {
+            return;
+        }
+
+        // check if this is an incomplete entity
+        try {
+            e.complete();
+        } catch (IncompleteEntityException exc) {
+            // not complete and link entity does not exist
             throw new NoSuchEntityException("No such entity '" + e.getEntityType() + "' with id " + e.getId().getValue());
         }
+
+        // complete with id -> create
+        pm.insert(e);
+        return;
     }
 
     public boolean entityExists(Entity e) {